% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheck_data.R
\name{errcheck_data}
\alias{errcheck_data}
\title{Error check a data matrix.}
\usage{
errcheck_data(X, calledby)
}
\arguments{
\item{X}{a matrix with counts or densities arranged in species by years. No NAs or negative values allowed,
constant species not allowed.}

\item{calledby}{the function calling this one}
}
\value{
\code{errcheck_data} returns nothing but throws and error if the inputs do not
meet the requirements
}
\description{
Error check a data matrix.
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
