% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheck_tsvreq.R
\name{errcheck_tsvreq}
\alias{errcheck_tsvreq}
\title{Error check inputs for the creator function for the \code{tsvreq} class.}
\usage{
errcheck_tsvreq(ts, com, comnull, tsvr, wts)
}
\arguments{
\item{ts}{Timescales, should be a numeric vector of nonnegative numbers}

\item{com}{Should be a vector of nonnegative numbers, equal to \code{comnull}*\code{tsvr}}

\item{comnull}{Should be a vector of nonnegative numbers}

\item{tsvr}{Should be a vector of nonnegative numbers}

\item{wts}{Should be a vector of nonnegative numbers}
}
\value{
\code{errcheck_tsvreq} returns nothing but throws and error if the inputs do not
meet the requirements of a \code{tsvreq} object
}
\description{
Error check inputs for the creator function for the \code{tsvreq} class.
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
