% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheck_vreq.R
\name{errcheck_vreq}
\alias{errcheck_vreq}
\title{Error check inputs for the creator function for the \code{vreq} class.}
\usage{
errcheck_vreq(com, comnull, vr)
}
\arguments{
\item{com}{Should be a positive number, equal to \code{comnull}*\code{vr}}

\item{comnull}{Should be a positive number}

\item{vr}{Should be a positive number}
}
\value{
\code{errcheck_vreq} returns nothing but throws and error if the inputs do not
meet the requirements of a \code{vreq} object
}
\description{
Error check inputs for the creator function for the \code{vreq} class.
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
