% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary_tsvr.R
\name{print.summary_tsvr}
\alias{print.summary_tsvr}
\title{Print method for \code{summary_tsvr} class}
\usage{
\method{print}{summary_tsvr}(x, ...)
}
\arguments{
\item{x}{A \code{summary_tsvr} object}

\item{...}{Not currently used. Included for argument consistency
with existing generics.}
}
\value{
\code{print.summary_tsvr} is called for its effect of
printing to the screen.
}
\description{
Print method for \code{summary_tsvr} class
}
\examples{
res<-vreq(2,1,2)
summary(res)

}
\seealso{
\code{\link{vreq_methods}}, \code{\link{vreq_classic_methods}}, 
\code{\link{vreq_LdM_methods}}, \code{\link{vreq_classic_ag_methods}}, 
\code{\link{tsvreq_methods}}, \code{\link{tsvreq_classic_methods}}, 
\code{browseVignettes("tsvr")}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
