% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvreq_classic_methods.R
\name{tsvreq__classic_methods}
\alias{tsvreq__classic_methods}
\title{Basic methods for the \code{tsvreq_classic} class}
\arguments{
\item{object, x, obj}{An object of class \code{tsvreq_classic}}

\item{newval}{A new value, for the \code{set_*} methods}

\item{...}{Not currently used. Included for argument consistency
with existing generics.}
}
\value{
\code{summary.tsvreq_classic} produces a summary of a \code{tsvreq_classic} object.
Methods \code{print.tsvreq_classic} and \code{plot.tsvreq_classic} are also available. 
For \code{tsvreq_classic} objects, 
\code{set_*} and \code{get_*} methods are available for all slots (see
the documentation for \code{tsvreq_classic} for a list). The \code{set_*} methods 
just throw an error, to prevent breaking the consistency between the 
slots of a \code{tsvreq_classic} object.
}
\description{
Set, get, summary, print and plot methods for the \code{tsvreq_classic} class.
}
\examples{
add later
 
}
\seealso{
\code{\link{tsvreq_classic}}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
