% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvreq_classic_methods.R
\name{tsvreq_classic_methods}
\alias{tsvreq_classic_methods}
\alias{summary.tsvreq_classic}
\alias{print.tsvreq_classic}
\alias{plot.tsvreq_classic}
\alias{set_ts.tsvreq_classic}
\alias{set_com.tsvreq_classic}
\alias{set_comnull.tsvreq_classic}
\alias{set_tsvr.tsvreq_classic}
\alias{set_wts.tsvreq_classic}
\alias{get_ts.tsvreq_classic}
\alias{get_com.tsvreq_classic}
\alias{get_comnull.tsvreq_classic}
\alias{get_tsvr.tsvreq_classic}
\alias{get_wts.tsvreq_classic}
\title{Basic methods for the \code{tsvreq_classic} class}
\usage{
\method{summary}{tsvreq_classic}(object, ...)

\method{print}{tsvreq_classic}(x, ...)

\method{plot}{tsvreq_classic}(x, filename = NA, ...)

\method{set_ts}{tsvreq_classic}(obj, newval)

\method{set_com}{tsvreq_classic}(obj, newval)

\method{set_comnull}{tsvreq_classic}(obj, newval)

\method{set_tsvr}{tsvreq_classic}(obj, newval)

\method{set_wts}{tsvreq_classic}(obj, newval)

\method{get_ts}{tsvreq_classic}(obj)

\method{get_com}{tsvreq_classic}(obj)

\method{get_comnull}{tsvreq_classic}(obj)

\method{get_tsvr}{tsvreq_classic}(obj)

\method{get_wts}{tsvreq_classic}(obj)
}
\arguments{
\item{object, x, obj}{An object of class \code{tsvreq_classic}}

\item{...}{Passed to plot. Not currently used for other methods, included there only for argument consistency
with existing generics.}

\item{filename}{A filename, no extension, could have a path. Used for saving a plot as a pdf. The default value NA causes the default plotting device to be used.}

\item{newval}{A new value, for the \code{set_*} methods}
}
\value{
\code{summary.tsvreq_classic} produces a summary of a \code{tsvreq_classic} object.
Methods \code{print.tsvreq_classic} and \code{plot.tsvreq_classic} are also available. 
For \code{tsvreq_classic} objects, 
\code{set_*} and \code{get_*} methods are available for all slots (see
the documentation for \code{tsvreq_classic} for a list). The \code{set_*} methods 
just throw an error, to prevent breaking the consistency between the 
slots of a \code{tsvreq_classic} object.
}
\description{
Set, get, summary, print and plot methods for the \code{tsvreq_classic} class.
}
\examples{
X<-matrix(runif(10*100),10,100)
res<-tsvreq_classic(X)
get_ts(res)
print(res)
summary(res)
 
}
\references{
Zhao et al, (In prep) Decomposition of the variance ratio illuminates timescale-specific
population and community variability.
}
\seealso{
\code{\link{tsvreq_classic}}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
