% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vreq_LdM_methods.R
\name{vreq_LdM_methods}
\alias{vreq_LdM_methods}
\alias{summary.vreq_LdM}
\alias{print.vreq_LdM}
\alias{set_com.vreq_LdM}
\alias{set_comnull.vreq_LdM}
\alias{set_vr.vreq_LdM}
\alias{get_com.vreq_LdM}
\alias{get_comnull.vreq_LdM}
\alias{get_vr.vreq_LdM}
\title{Basic methods for the \code{vreq_LdM} class}
\usage{
\method{summary}{vreq_LdM}(object, ...)

\method{print}{vreq_LdM}(x, ...)

\method{set_com}{vreq_LdM}(obj, newval)

\method{set_comnull}{vreq_LdM}(obj, newval)

\method{set_vr}{vreq_LdM}(obj, newval)

\method{get_com}{vreq_LdM}(obj)

\method{get_comnull}{vreq_LdM}(obj)

\method{get_vr}{vreq_LdM}(obj)
}
\arguments{
\item{object, x, obj}{An object of class \code{vreq_LdM}}

\item{...}{Not currently used. Included for argument consistency
with existing generics.}

\item{newval}{A new value, for the \code{set_*} methods}
}
\value{
\code{summary.vreq_LdM} produces a summary of a \code{vreq_LdM} object.
A \code{print.vreq_LdM} method is also available. For \code{vreq_LdM} objects, 
\code{set_*} and \code{get_*} methods are available for all slots (see
the documentation for \code{vreq_LdM} for a list). The \code{set_*} methods 
just throw an error, to prevent breaking the consistency between the 
slots of a \code{vreq_LdM} object.
}
\description{
Set, get, summary, and print methods for the \code{vreq_LdM} class.
}
\examples{
add later
 
}
\seealso{
\code{\link{vreq_LdM}}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
