% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vreq_methods.R
\name{vreq_methods}
\alias{vreq_methods}
\alias{summary.vreq}
\alias{print.vreq}
\alias{set_com.vreq}
\alias{set_comnull.vreq}
\alias{set_vr.vreq}
\alias{get_com.vreq}
\alias{get_comnull.vreq}
\alias{get_vr.vreq}
\title{Basic methods for the \code{vreq} class}
\usage{
\method{summary}{vreq}(object, ...)

\method{print}{vreq}(x, ...)

\method{set_com}{vreq}(obj, newval)

\method{set_comnull}{vreq}(obj, newval)

\method{set_vr}{vreq}(obj, newval)

\method{get_com}{vreq}(obj)

\method{get_comnull}{vreq}(obj)

\method{get_vr}{vreq}(obj)
}
\arguments{
\item{object, x, obj}{An object of class \code{vreq}}

\item{...}{Not currently used. Included for argument consistency
with existing generics.}

\item{newval}{A new value, for the \code{set_*} methods}
}
\value{
\code{summary.vreq} produces a summary of a \code{vreq} object.
A \code{print.vreq} method is also available. For \code{vreq} objects, 
\code{set_*} and \code{get_*} methods are available for all slots (see
the documentation for \code{vreq} for a list). The \code{set_*} methods 
just throw an error, to prevent breaking the consistency between the 
slots of a \code{vreq} object.
}
\description{
Set, get, summary, and print methods for the \code{vreq} class.
}
\examples{
res<-vreq(com=2,comnull=1,vr=2)
print(res)
 
}
\seealso{
\code{\link{vreq}}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
