% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wts.R
\name{wts}
\alias{wts}
\title{Compute the weights (wts)}
\usage{
wts(X)
}
\arguments{
\item{X}{a matrix with counts or densities arranged in species by time step}
}
\value{
\code{wts} returns an object of class list consisting of
\item{frequency}{a vector from 0 to 1 (not include 0 and 1)}
\item{wts}{a vector of wts}
}
\description{
This function is used to compute weights (wts)
}
\examples{
X<-matrix(runif(200,1,100), 10, 20)
ans<-wts(X)

}
\references{
<Lei paper>
}
\author{
Lei Zhao, \email{lei.zhao@cau.edu.cn}; Daniel Reuman, \email{reuman@ku.edu}
}
