% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggts.R
\name{aggts}
\alias{aggts}
\title{Aggregates a \code{tsvreq_classic} object across a set of timescales; also the constructor function
for class \code{vreq_classic_ag}.}
\usage{
aggts(obj, ts)
}
\arguments{
\item{obj}{A \code{tsvreq_classic} object}

\item{ts}{The timescales to aggregate over}
}
\value{
\code{aggts} returns an object of class \code{vreq_classic_ag}. Slots are:
\item{com}{the timescale-aggregated value of CVcom2}
\item{comnull}{the timescale-aggregated value of CVcomip2}
\item{vr}{the timescale-aggregated value of the classic variance ratio}
\item{ts}{the timescales over which aggregation was performed}
}
\description{
All the components of a \code{tsvreq_classic} object can be aggregated across an arbitrary 
set of timescales, producing a new variance ratio equation - this function performs that
aggregation. The function returns a \code{vreq_classic_ag} object, and is the constructor 
function of that class. The 
\code{vreq_classic_ag} class has slots \code{com}, \code{comnull}, \code{vr}, which are
the same as a \code{vreq} object, but also has slot \code{ts}, which is the timescales
over which aggregation was performed to get the object. The class inherits from \code{vreq},
which inherits from \code{list}.
}
\examples{
X<-matrix(runif(10*100),10,100)
h<-tsvreq_classic(X)
res<-aggts(h,h$ts[h$ts>4])

}
\references{
<add Lei's paper>
}
\seealso{
\code{\link{tsvreq_classic}}, \code{\link{vreq_classic_ag_methods}}, 
\code{browseVignettes("tsvr")}
}
\author{
Shaopeng Wang, \email{shaopeng.wang@pku.edu.cn}; Lei Zhao, \email{lei.zhao@cau.edu.cn}; Daniel Reuman, \email{reuman@ku.edu}
}
