% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cospect.R
\name{cospect}
\alias{cospect}
\title{Calculate the cospectrum between all pairs (including pairs with itself)}
\usage{
cospect(X)
}
\arguments{
\item{X}{a matrix with counts or densities arranged in species by time step.}
}
\value{
\code{cospect} return a list with elements
\item{frequency}{a vector from 0 to 1 of the frequencies used}
\item{cospectrum}{a 3D array, with cospectrum range in species by species by frequency}
}
\description{
This function is used to calculate the cospectrum.
}
\examples{
X<-matrix(runif(200,1,100), 10, 20)
rownames(X)<-letters[1:10]
colnames(X)<-1991:2010
ans<-cospect(X)

}
\author{
Lei Zhao, \email{lei.zhao@cau.edu.cn}; Daniel Reuman, \email{reuman@ku.edu}
}
\keyword{internal}
