% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary_tsvr.R
\name{print.summary_tsvr}
\alias{print.summary_tsvr}
\title{Print method for \code{summary_tsvr} class}
\usage{
\method{print}{summary_tsvr}(x, ...)
}
\arguments{
\item{x}{A \code{summary_tsvr} object}

\item{...}{Not currently used. Included for argument consistency
with existing generics.}
}
\value{
\code{print.summary_tsvr} is called for its effect of
printing to the screen.
}
\description{
Print method for \code{summary_tsvr} class
}
\examples{
#Fill in later

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
