% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvreq_classic.R
\name{tsvreq_classic}
\alias{tsvreq_classic}
\title{Creator function for the \code{tsvreq_classic} S3 class}
\usage{
tsvreq_classic(X)
}
\arguments{
\item{X}{a matrix with counts or densities arranged in species by years}
}
\value{
\code{tsvreq_classic} returns a \code{tsvreq_classic} object. Slots are:
\item{ts} a vector of timescales
\item{com} a timescale-specific decomposition of CVcom2
\item{comnull} a timescale-specific decomposition of CVcomip2
\item{tsvr} a timescale-specific version of the classic variance ratio
\item{wts} a vector of weights, same length as all the above
}
\description{
The \code{tsvreq_classic} (timescale-specific variance ratio equation, classic variance
ratio) class is for storing functional equations based on a timescale specific version of
the classic variance ratio. Inherits from \code{tsvreq}, which inherits from \code{list}.
}
\examples{
#add later

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
