% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vreq_classic.R
\name{vreq_classic}
\alias{vreq_classic}
\title{Creator function of \code{vreq_classic} S3 class}
\usage{
vreq_classic(X)
}
\arguments{
\item{X}{A matrix with counts or densities arranged in species by years}
}
\value{
\code{vreq_classic} returns a \code{vreq_classic} object. Slots are:
\item{com} the squared community CV, CVcom2
\item{comnull} CVcomip2
\item{vr} the classic variance ratio
}
\description{
The \code{vreq_classic} (variance ratio equation, classic variance ratio) class is for 
storing equations based on the classic variance ratio. Inherits from the \code{vreq} class,
which inherits from the \code{list} class.
}
\examples{
#add later

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
